+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file documents the changes made to the formatter with each release.
 
 ## Release 0.22.2 (2026-07-22)
 
+### Added
+
+- Handle Tree-sitter parse errors gracefully, preserve declarations with parse errors while formatting valid code around them (#279)
+
 ### Fixed
 
 - Fix long ternaries without pre-existing backslashes or delimiters formatting in a way that GDScript doesn't parse (the formatter will now add parentheses around the expression and indent it) (#293)